+++ api/src/main/java/com/alibaba/nacos/api/naming/pojo/Instance.java
@@ -17,6 +17,7 @@ package com.alibaba.nacos.api.naming.pojo;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.annotation.JSONField;
+import com.alibaba.nacos.api.common.Constants;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -67,7 +68,7 @@ public class Instance {
     /**
      * User extended attributes
      */
-    private Map<String, String> metadata = new HashMap<>();
+    private Map<String, String> metadata = new HashMap<String, String>();
 
     public String getInstanceId() {
         return instanceId;
@@ -75,6 +76,14 @@ public class Instance {
 
     public void setInstanceId(String instanceId) {
         this.instanceId = instanceId;
+    }
+
+    public String serviceName() {
+        String[] infos = instanceId.split(Constants.NAMING_INSTANCE_ID_SPLITTER);
+        if (infos.length < Constants.NAMING_INSTANCE_ID_SEG_COUNT) {
+            return null;
+        }
+        return infos[Constants.NAMING_INSTANCE_ID_SEG_COUNT - 1];
     }
 
     public String getIp() {
@@ -165,5 +174,5 @@ public class Instance {
     private static boolean strEquals(String str1, String str2) {
         return str1 == null ? str2 == null : str1.equals(str2);
     }
-    
+
 }