+++ test/src/test/java/com/alibaba/nacos/test/naming/Params.java
@@ -27,7 +27,7 @@ public class Params {
 
     public static Params newParams() {
         Params params = new Params();
-        params.paramMap = new LinkedMultiValueMap<>();
+        params.paramMap = new LinkedMultiValueMap<String, String>();
         return params;
     }