+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.0 (Sept 17, 2018)
+
+#### FEATURES:
+
+* separate nacos-api from nacos client implementation
+* Cluster node health check abnormality
+* Stand-alone mode gets the change history list exception
+* Pulling does not exist configuration print io exception
+* Optimized log framework
+* Service Discovery: Client support getting server status.
+* Service Discovery: Client support get all service names of server.
+* Service Discovery: Client support get all subscribed services.
+
+
+#### IMPROVEMENTS:
+
+#### BUG FIXES:
+
+#### BREAKING CHANGES:
+
+
+
 ## 0.1.0 (July 18, 2018)
 
 #### FEATURES: